+++ RandomSkunk.Results.Http/ProblemDetails.cs
@@ -1,0 +1,24 @@
+namespace RandomSkunk.Results.Http;
+
+// https://github.com/dotnet/aspnetcore/blob/0ee742c53f2669fd7233df6da89db5e8ab944585/src/Http/Http.Extensions/src/ProblemDetails.cs
+[JsonConverter(typeof(ProblemDetailsJsonConverter))]
+internal class ProblemDetails
+{
+    [JsonPropertyName("type")]
+    public string? Type { get; set; }
+
+    [JsonPropertyName("title")]
+    public string? Title { get; set; }
+
+    [JsonPropertyName("status")]
+    public int? Status { get; set; }
+
+    [JsonPropertyName("detail")]
+    public string? Detail { get; set; }
+
+    [JsonPropertyName("instance")]
+    public string? Instance { get; set; }
+
+    [JsonExtensionData]
+    public IDictionary<string, object?> Extensions { get; } = new Dictionary<string, object?>(StringComparer.Ordinal);
+}